fix wrong input ::floor

+++ libvis/src/libvis/cuda/pixel_corner_projector.cuh
@@ -147,8 +147,8 @@ struct PixelCornerProjector_ {
     
     float fc = (undistorted_nxy.x - min_nx) * ((resolution_x - 1) / (max_nx - min_nx));
     float fr = (undistorted_nxy.y - min_ny) * ((resolution_y - 1) / (max_ny - min_ny));
-    const int row = ::floor(fr);
-    const int col = ::floor(fc);
+    const int row = ::floorf(fr);
+    const int col = ::floorf(fc);
     float r_frac = fr - row;
     float c_frac = fc - col;
